Atualizando

+++ ProgramacaoDeComputadores/Lista01Funcao/ativ_003.c
@@ -6,15 +6,15 @@
 #include <locale.h>
 #include <math.h>
 
-int distancia() {
-  int x1, x2, y1, y2;
+float distancia() {
+  float x1, x2, y1, y2;
   float distancia;
 
   printf("Digite as coordenadas do primeiro ponto:");
-  scanf("%d %d", &x1, &y1);
+  scanf("%f %f", &x1, &y1);
 
   printf("Digite as coordenadas do segundo ponto:");
-  scanf("%d %d", &x2, &y2);
+  scanf("%f %f", &x2, &y2);
 
   distancia = sqrt (pow(x2-x1,2)+pow(y2-y1,2));
